refactor: move test related files to new `tests` directory
